+++ commercetools/commercetools-sdk-java-api/src/test/java-generated/com/commercetools/api/client/resource/ByProjectKeyRecurringOrdersKeyByKeyTest.java
@@ -87,7 +87,36 @@ public class ByProjectKeyRecurringOrdersKeyByKeyTest {
                         .recurringOrders()
                         .withKey("test_key")
                         .post(com.commercetools.api.models.recurring_order.RecurringOrderUpdate.of())
-                        .createHttpRequest(), "post", "test_projectKey/recurring-orders/key=test_key", } };
+                        .createHttpRequest(), "post", "test_projectKey/recurring-orders/key=test_key", },
+                new Object[] {
+                        apiRoot.withProjectKey("test_projectKey")
+                                .recurringOrders()
+                                .withKey("test_key")
+                                .delete()
+                                .withDataErasure(true)
+                                .createHttpRequest(),
+                        "delete", "test_projectKey/recurring-orders/key=test_key?dataErasure=true", },
+                new Object[] {
+                        apiRoot.withProjectKey("test_projectKey")
+                                .recurringOrders()
+                                .withKey("test_key")
+                                .delete()
+                                .withVersion(2)
+                                .createHttpRequest(),
+                        "delete", "test_projectKey/recurring-orders/key=test_key?version=2", },
+                new Object[] {
+                        apiRoot.withProjectKey("test_projectKey")
+                                .recurringOrders()
+                                .withKey("test_key")
+                                .delete()
+                                .withExpand("expand")
+                                .createHttpRequest(),
+                        "delete", "test_projectKey/recurring-orders/key=test_key?expand=expand", },
+                new Object[] { apiRoot.withProjectKey("test_projectKey")
+                        .recurringOrders()
+                        .withKey("test_key")
+                        .delete()
+                        .createHttpRequest(), "delete", "test_projectKey/recurring-orders/key=test_key", } };
     }
 
     public static Object[][] executeMethodParameters() {
@@ -108,6 +137,23 @@ public class ByProjectKeyRecurringOrdersKeyByKeyTest {
                 new Object[] { apiRoot.withProjectKey("test_projectKey")
                         .recurringOrders()
                         .withKey("test_key")
-                        .post(com.commercetools.api.models.recurring_order.RecurringOrderUpdate.of()), } };
+                        .post(com.commercetools.api.models.recurring_order.RecurringOrderUpdate.of()), },
+                new Object[] { apiRoot.withProjectKey("test_projectKey")
+                        .recurringOrders()
+                        .withKey("test_key")
+                        .delete()
+                        .withDataErasure(true), },
+                new Object[] { apiRoot.withProjectKey("test_projectKey")
+                        .recurringOrders()
+                        .withKey("test_key")
+                        .delete()
+                        .withVersion(2), },
+                new Object[] { apiRoot.withProjectKey("test_projectKey")
+                        .recurringOrders()
+                        .withKey("test_key")
+                        .delete()
+                        .withExpand("expand"), },
+                new Object[] {
+                        apiRoot.withProjectKey("test_projectKey").recurringOrders().withKey("test_key").delete(), } };
     }
 }