+++ changes.md
@@ -3,6 +3,8 @@
 <details>
 <summary>Added Type(s)</summary>
 
+- added type `RecurringOrderDeletedMessage`
+- added type `RecurringOrderDeletedMessagePayload`
 - added type `ProductSearchFacetResultStats`
 - added type `ProductSearchFacetStatsExpression`
 - added type `ProductSearchFacetStatsValue`
@@ -13,5 +15,13 @@
 <summary>Removed Type(s)</summary>
 
 - :warning: removed type `RecurringOrderMessagePayload`
+</details>
+
+
+<details>
+<summary>Added Method(s)</summary>
+
+- added method `apiRoot.withProjectKey().recurringOrders().withId().delete()`
+- added method `apiRoot.withProjectKey().recurringOrders().withKey().delete()`
 </details>